+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 #### Improvements
 
-  - Added the `pgtrigger.CondValues` trigger utility for writing conditional statement-level triggers by [@wesleykendall](https://github.com/wesleykendall) in [#204](https://github.com/AmbitionEng/django-pgtrigger/pull/204).
+  - Added the `pgtrigger.Composer` trigger utility for writing conditional statement-level triggers by [@wesleykendall](https://github.com/wesleykendall) in [#204](https://github.com/AmbitionEng/django-pgtrigger/pull/204).
 
 ## 4.14.0 (2025-04-18)